# doc/api-docs-smart/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PinShop Server",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
...
    "/api/v1/product/flash-list": {
      "get": {
        "summary": "获取秒杀商品流\n\n<p>\n返回当前可抢购的秒杀商品列表。\n前端用于首页展示秒杀商品卡片。\n</p>",
        "deprecated": false,
        "description": "建议前端每30秒刷新一次列表，获取最新库存状态",
        "tags": [
          "ProductController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/4dabcffbb9f9a634ef699122e06c5027"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": [\n    {\n      \"skuId\": 0,\n      \"title\": \"\",\n      \"imgUrl\": \"\",\n      \"flashPrice\": 0,\n      \"originalPrice\": 0,\n      \"stockPercent\": 0,\n      \"status\": 0\n    }\n  ],\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-product-flash-list-GET",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "时间段（可选，如\"08:00\"、\"12:00\"、\"20:00\"）",
            "required": false,
            "name": "timeSlot"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      }
    },
    "/api/v1/product/detail/{skuId}": {
      "get": {
        "summary": "获取商品详情\n\n<p>\n返回指定SKU的商品详细信息。\n</p>",
        "deprecated": false,
        "tags": [
          "ProductController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/388eafe398d013c10e7a03d03dfe963f"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {\n    \"skuId\": 0,\n    \"title\": \"\",\n    \"imgUrl\": \"\",\n    \"flashPrice\": 0,\n    \"originalPrice\": 0,\n    \"stockPercent\": 0,\n    \"status\": 0\n  },\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-product-detail-{skuId}-GET",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "path",
            "description": "商品SKU编号",
            "required": true,
            "example": "0",
            "name": "skuId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      }
    },
    "/api/v1/product/categories": {
      "get": {
        "summary": "获取商品分类\n\n<p>\n返回商品分类列表，用于分类导航。\n</p>",
        "deprecated": false,
        "tags": [
          "ProductController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/7fb29907321bcf259e566af25a78b727"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": [\n    {\n      \"categoryId\": 0,\n      \"categoryName\": \"\",\n      \"categoryIcon\": \"\"\n    }\n  ],\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-product-categories-GET",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      }
    },
    "/api/v1/product/search": {
      "get": {
        "summary": "搜索商品\n\n<p>\n根据关键词搜索商品。\n</p>",
        "deprecated": false,
        "tags": [
          "ProductController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/cac4c6a3a7ba7166ee737807946f1059"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {\n    \"total\": 0,\n    \"products\": [\n      {\n        \"skuId\": 0,\n        \"title\": \"\",\n        \"imgUrl\": \"\",\n        \"flashPrice\": 0,\n        \"originalPrice\": 0,\n        \"stockPercent\": 0,\n        \"status\": 0\n      }\n    ]\n  },\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-product-search-GET",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "搜索关键词",
            "required": true,
            "name": "keyword"
          },
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "description": "页码（默认1）",
            "required": true,
            "example": "\"1\"",
            "name": "pageNum"
          },
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "description": "每页大小（默认20）",
            "required": true,
            "example": "\"20\"",
            "name": "pageSize"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      }
    },
    "/api/v1/product/hot": {
      "get": {
        "summary": "获取热门商品\n\n<p>\n返回销量最高的热门商品列表。\n</p>",
        "deprecated": false,
        "tags": [
          "ProductController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/4dabcffbb9f9a634ef699122e06c5027"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": [\n    {\n      \"skuId\": 0,\n      \"title\": \"\",\n      \"imgUrl\": \"\",\n      \"flashPrice\": 0,\n      \"originalPrice\": 0,\n      \"stockPercent\": 0,\n      \"status\": 0\n    }\n  ],\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-product-hot-GET",
        "parameters": [
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "description": "返回数量（默认10）",
            "required": true,
            "example": "\"10\"",
            "name": "limit"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ed754232fb2bdf590bcc26aba799beae": {
        "properties": {
          "skuId": {
            "format": "int64",
            "description": "商品SKU编号",
            "type": "number",
            "example": "0"
          },
          "title": {
            "format": "string",
            "description": "商品标题",
            "type": "string"
          },
          "imgUrl": {
            "format": "string",
            "description": "商品主图URL",
            "type": "string"
          },
          "flashPrice": {
            "description": "秒杀价格",
            "example": "0"
          },
          "originalPrice": {
            "description": "原价（划线价）",
            "example": "0"
          },
          "stockPercent": {
            "format": "int32",
            "description": "剩余库存百分比（0-100，用于前端进度条显示）",
            "type": "integer",
            "example": "0"
          },
          "status": {
            "format": "int32",
            "description": "状态（1:抢购中 2:已抢光）",
            "type": "integer",
            "example": "0"
          }
        }
      },
      "4dabcffbb9f9a634ef699122e06c5027": {
        "properties": {
          "code": {
            "format": "int32",
            "description": "业务状态码",
            "type": "integer",
            "example": "200"
          },
          "message": {
            "format": "string",
            "description": "响应消息",
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "业务数据",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/e8105cc96d70913bfb4521803969636b"
            }
          },
          "timestamp": {
            "format": "int64",
            "description": "响应时间戳(毫秒)",
            "type": "number",
            "example": "1709823456789"
          }
        }
      },
...
      "83c6f2daad4ff1df851a2cc1d9595e9d": {
        "properties": {
          "total": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": "0"
          },
          "products": {
            "description": "No comments found.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ed754232fb2bdf590bcc26aba799beae"
            }
          }
        }
      },
...
      "cac4c6a3a7ba7166ee737807946f1059": {
        "properties": {
          "code": {
            "format": "int32",
            "description": "业务状态码",
            "type": "integer",
            "example": "200"
          },
          "message": {
            "format": "string",
            "description": "响应消息",
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "业务数据(object)",
            "$ref": "#/components/schemas/83c6f2daad4ff1df851a2cc1d9595e9d"
          },
          "timestamp": {
            "format": "int64",
            "description": "响应时间戳(毫秒)",
            "type": "number",
            "example": "1709823456789"
          }
        }
      },
...
      "288c302109dc389aea3311b8d9a9d8be": {
        "properties": {
          "categoryId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "0"
          },
          "categoryName": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "categoryIcon": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          }
        }
      },
...
      "7fb29907321bcf259e566af25a78b727": {
        "properties": {
          "code": {
            "format": "int32",
            "description": "业务状态码",
            "type": "integer",
            "example": "200"
          },
          "message": {
            "format": "string",
            "description": "响应消息",
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "业务数据",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/288c302109dc389aea3311b8d9a9d8be"
            }
          },
          "timestamp": {
            "format": "int64",
            "description": "响应时间戳(毫秒)",
            "type": "number",
            "example": "1709823456789"
          }
        }
      },
...
      "e8105cc96d70913bfb4521803969636b": {
        "properties": {
          "skuId": {
            "format": "int64",
            "description": "商品SKU编号",
            "type": "number",
            "example": "0"
          },
          "title": {
            "format": "string",
            "description": "商品标题",
            "type": "string"
          },
          "imgUrl": {
            "format": "string",
            "description": "商品主图URL",
            "type": "string"
          },
          "flashPrice": {
            "description": "秒杀价格",
            "example": "0"
          },
          "originalPrice": {
            "description": "原价（划线价）",
            "example": "0"
          },
          "stockPercent": {
            "format": "int32",
            "description": "剩余库存百分比（0-100，用于前端进度条显示）",
            "type": "integer",
            "example": "0"
          },
          "status": {
            "format": "int32",
            "description": "状态（1:抢购中 2:已抢光）",
            "type": "integer",
            "example": "0"
          }
        }
      },
      "388eafe398d013c10e7a03d03dfe963f": {
        "properties": {
          "code": {
            "format": "int32",
            "description": "业务状态码",
            "type": "integer",
            "example": "200"
          },
          "message": {
            "format": "string",
            "description": "响应消息",
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "业务数据(object)",
            "$ref": "#/components/schemas/e8105cc96d70913bfb4521803969636b"
          },
          "timestamp": {
            "format": "int64",
            "description": "响应时间戳(毫秒)",
            "type": "number",
            "example": "1709823456789"
          }
        }
      },
...
      "da0c932b0e6d3b90859201037a7929ee": {
        "properties": {
          "code": {
            "format": "int32",
            "description": "业务状态码",
            "type": "integer",
            "example": "200"
          },
          "message": {
            "format": "string",
            "description": "响应消息",
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "业务数据(object)",
            "type": "object"
          },
          "timestamp": {
            "format": "int64",
            "description": "响应时间戳(毫秒)",
            "type": "number",
            "example": "1709823456789"
          }
        }
      },
...
    }
  }
}